requirements.txt
absl-py==2.4.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1764961372591/work
certifi @ file:///C:/Users/task_176765902528894/croot/certifi_1767659216804/work/certifi
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832794697/work
charset-normalizer @ file:///C:/miniconda3/conda-bld/charset-normalizer_1761744944645/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1769745019/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
filelock @ file:///C:/Users/task_176943980686702/croot/filelock_1769440565956/work
fsspec==2026.1.0
gmpy2 @ file:///C:/miniconda3/conda-bld/gmpy2_1764748105976/work
grpcio==1.78.0
idna @ file:///C:/miniconda3/conda-bld/idna_1761912044311/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1738273805233/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1761567966/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711254/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1767954655/work
jupyter_core @ file:///D:/bld/bld/rattler-build_jupyter_core_1760643889/work
lazy_loader @ file:///home/conda/feedstock_root/build_artifacts/lazy-loader_1733636780672/work
Markdown==3.10.2
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
monai @ file:///home/conda/feedstock_root/build_artifacts/monai_1769585314056/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
nibabel @ file:///home/conda/feedstock_root/build_artifacts/nibabel_1734022767841/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp310-cp310-win_amd64.whl#sha256=e824ec3b279f4c5207736c83ac2f2dc79b489e08501a4ea9e3fc178c45db289b
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas==2.3.3
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1764950726/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
protobuf==6.33.5
psutil @ file:///D:/bld/bld/rattler-build_psutil_1769678167/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496059965/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753030270/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz==2025.2
PyWavelets==1.8.0
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487147/work
PyYAML @ file:///C:/miniconda3/conda-bld/pyyaml_1763465513320/work
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387021/work
requests @ file:///C:/miniconda3/conda-bld/requests_1762359612241/work
scikit-image @ file:///D:/bld/scikit-image_1757196766804/work
scipy @ file:///C:/bld/scipy-split_1739790833536/work/dist/scipy-1.15.2-cp310-cp310-win_amd64.whl#sha256=4748853a0c723cda1fa2bdead847ba7c2443bde9cec3436e3e7243cfa54b55ff
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tifffile @ file:///C:/b/abs_8aikas7x_4/croot/tifffile_1741164551582/work
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado @ file:///D:/bld/tornado_1765836387919/work
tqdm @ file:///D:/bld/bld/rattler-build_tqdm_1769864679/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
tzdata==2025.3
urllib3 @ file:///C:/Users/task_176781036207501/croot/urllib3_1767810484846/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1769858610910/work
Werkzeug==3.1.5
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746276406/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_zipp_1764460141/work
